+++ antifraud/src/main/java/com/bank/antifraud/repository/AuditRepository.java
@@ -4,10 +4,12 @@ import com.bank.antifraud.entity.AuditEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 /**
  * Слой доступа к таблице Audit в БД
  * @author Makariy Petrov
  */
 @Repository
-public interface AuditRepository extends JpaRepository<AuditEntity, Long> {
+public interface AuditRepository extends JpaRepository<AuditEntity, BigInteger> {
 }
